+++ src/Tsfm.Forecasting.TimesFm/TimesFmModules.cs
@@ -1,10 +1,9 @@
-using TorchSharp;
+using System;
+using System.Linq;
 using TorchSharp.Modules;
 
 using static TorchSharp.torch;
 using static TorchSharp.torch.nn;
-
-using Tsfm.Forecasting;
 
 namespace Tsfm.Forecasting.TimesFm;